MiServiceIBinder.java: Se cambio el ID del Alerta Externa harcodeado por la variable estatica ID_SONIDO_ALERTA_EXTERNA SonidoAlertaActivity.java: Se modifico para que herede de Menu y pueda inflar el toolbar como los demas Activity. Se cambia el titulo del Alerta Activity y se hacen adaptaciones para que pueda heredar de Menu SonidosEdicionActivity.java: Se agrega una validacion para que antes de borrar la imagen relacionada a un sonido, valide que exista. Antes pinchaba con "Alerta Externa" que no tenia una imagen por defecto guardada.

+++ NotSound/app/src/main/java/com/example/gabys/notsound/SonidosEdicionActivity.java
@@ -196,8 +196,10 @@ public class SonidosEdicionActivity extends Menu {
 
         if (itemSeleccionado != SonidosActivity.SONIDO_NUEVO) {
 
-            File file = new File(sonidos.getSonidoByPosition(itemSeleccionado).getRutaFoto());
-            file.delete();
+            if (sonidos.getSonidoByPosition(itemSeleccionado).getRutaFoto() != null){
+                File file = new File(sonidos.getSonidoByPosition(itemSeleccionado).getRutaFoto());
+                file.delete();
+            }
 
             sonidos.setSonido(itemSeleccionado,(new Sonido(IDSonido,nombreSonido,rutaFoto,estaHabilitado)));
             sonidos.saveSonidos(getApplicationContext());
@@ -210,7 +212,6 @@ public class SonidosEdicionActivity extends Menu {
         }
 
         saveImage(imagen, rutaFoto);
-
         this.onBackPressed();
     }
